"11:55am. https://youtu.be/REeaT2mWj6Y Inconvenient truths about sqrt(2) | Real numbers and limits Math Foundations 80 | N J Wildberger
Ah, I did not know that he had a video about `sqrt(2)`. Based on the description, we seem to be in a complete agreement. I'll watch this video then during the break.

12:45pm. https://youtu.be/REeaT2mWj6Y?t=1576

This particular bit of math is completely new to me.

I am going to swap the link in `[some kinds of reasoning](https://www.homeschoolmath.net/teaching/proof_square_root_2_irrational.php)` with this lecture by Wildberger. Wildberger is great. I should listen to whatever he has to say. I wanted to scorched earth that entire line of reasoning, but he cut to the root in a way I could not.

1:15pm. Where was I? I am done with both breakfast and that lecture by Wildberger. Since I did not get a chance to do so already, let me read Tog and see if the Satanophany raw is out.

After that I think I will go for a lecture by Bowling.

4:50pm. I didn't think I would end up doing housework for 3.5 hours since that last entry. Damn, I am tired. Maybe it is good that this happened today just as I was taking a break.

5pm. Let me take a look at that talk by Bowling.

6:10pm. Ah, screw that talk. I am done with lunch and I want to rest.

6:15pm. Tomorrow as per plan, I will read the no-regret learning book."