"8:50am. Another day is here. It seems the winter preparations are starting in earnest, so I will have less time to study than usual until that is done.
Regardless, today's plan will be to get started with the `Prediction, Learning, and Games` (PLG) book. I want to track down the evolutionary origin of strategy averaging which should hopefully give me insight into its nature. It is true that I can avoid having to understand it by using the + variants of the algorithms, but I want to be thorough in my studies.

9:05am. Let me start.

9:50am. It seems I will be doing housework the whole day today, so I'll have even less time to study than I expected. I guess I'll have breakfast early then.

10:25am. 22/403. Uf, the intro fooled me. This book is far more mathematical than I expected it would be. But the very first equation is pretty much exactly strategy averaging. No doubt what I am looking for. Let me do this for a few more minutes and then I will go do those chores.

10:40am. I have to go. I'll resume this later."