"10:05am. __fmease__: "So irrational numbers don't exist because they are not computable? In the field of computer science, computability (and time & space complexity) is one of the most important concepts obviously. It's worth pursuing new algorithms/data structures. But should we limit ourselves to computability in the field of mathematics as well? I don't think so."
me: "Why not?

I agree that there is a tradeoff there. Making logical consistency a principle of mathematics (and not just programming) would necessitate a radical rethink of various areas. This is a practical burden. It is also a mental burden. It is a bit like selling a stock at a lower price than one bought it - it feels like you take an immediate loss, whereas if you hold on there is always hope that it might go up and you'll be able to sell it at a profit.

But the fact is, as soon the stock went down, the loss is real even if you do not sell it. And if you accept the loss, you can free up the remaining capital to invest in more promising ventures.

Mathematics, or more precisely, the mathematical community might be beyond base monetary concerns, but the fact is there are absolutely such things as cognitive budgets. You, me, the whole community have only a limited amount of focus that can be allocated to do research in areas of interest. We might not dabble in stocks, but we constantly do credit assignment throughout our life to things that are important.

By internalizing a principle you narrow you field of search. You lose the ability to do some of the things you could do before, but you also gain the option to go further in areas you might not have thought possible.

Internalizing the right principles is important. Without principles, there would be nothing but the brute forcing of infinities."

10:20am. I've been writing that post. I really need to get to work on that reply to Lanctot. I saw the email by him at 11pm yesterday, but I was too tired to compose a reply.

...Actually rather than writing rant, let me just thank him for the references and I'll go into rant mode when I actually finish the studies. Hopefully he would not mind me pasting these into a public journal.

Marc Lanctot: "Mi Marko,

Cool. Here are a few replies.

On Sun, Sep 29, 2019 at 8:49 AM Marko Grdinic <mrakgr@gmail.com> wrote:
> It took me a while to figure out why in the regret matching equation, only the probability of the opposing player is used.

Yes, exactly. You want to think of it as a measure of regret given that player i plays to reach I with probability 1. So the "counterfactual" part of the regret is only taking into account the reach probability of the other players (since the reach probability of the player to play at I is equal to 1). This is an interesting way to define values (it's fundamentally different than the way values are defined in reinforcement learning -- see Section 3.2 of https://arxiv.org/abs/1810.09026)

> So for updating regrets, what CFR is doing is assuming that the current player is using an uniform strategy for exploration - at least in the preceding nodes.

No that's not quite right (but I see why you think of it that way). The nodes visited are as if they were sampled from a uniform distribution over all states.. but that's a sampling distribution, that's not the same as the actual current strategy used by player i being uniform.

When computing counterfactual values and counterfactual regret in vanilla CFR, the sampling parts do not come in since the values are not estimates, they are exact. The sampling distributions are more relevant in the Monte Carlo variants.

> Now that just leaves figuring out the strategy averaging.

> Considering what I wrote previously, strategy averaging assumes the the current player is following the strategy based off the current regrets, but that the opponent is using an uniform strategy. I guess not considering opponent probabilities does make sense on toy games like rock paper scissors, but going beyond that is an open problem for me. The same goes for why strategy averaging works in general.

The averaging of a strategy for player i is completely independent of the opponent strategy. The weighting by the reach probably is to make sure that the average strategy computed in extensive-form is the same as the equivalent exponentially larger normal form game. This is an application of Kuhn's theorem: every extensive-form game has an equivalent (exponentially large) normal form representation (see page 134 of http://www.masfoundations.org/mas.pdf). The average strategy is being computed in a way that ensures it is equivalent to the average in this normal form equivalent.

The construction of the sequence-form might help with this as well (Chapter 5 of that same book), of the XFP paper might be also a good reference: http://proceedings.mlr.press/v37/heinrich15.pdf . The best reference was Gintis's Game Theory Evolving, but they unfortunately removed it when they moved to second edition. I'll try to see if I can find those few pages that explain it, as there is currently no good reference."

Marc Lanctot: "Here's some overview on behavior vs mixed strategies: https://www.youtube.com/watch?v=tT0E7PaDVck"

10:40am. I've made a brief reply and asked him about the softmax and exponential averaging.

At any rate, let me start going through the references.

Let me read the first two chapter of the PRG book and then I will take a look at the references.

10:55am. Ok, ok, focus me. Stop browsing social media.

The book!

11:25am. 27/403. Holy shit. This book goes from a simple example in the first few pages, to the absolute arcane short afterwards. I can't follow all these sums, exponentials, Greek letters and inequalities. Even figuring just a single thing here feels like it could take me days. Change of plans, I will skim the book.

12pm. https://www.youtube.com/watch?v=49PHeGZzttM
"Trading without Regret" by Dr. Michael Kearns

I am going to drop the book as it is too complicated. It is a prime exhibit of 'needs to be put into a proof assistant'. It is way beyond what I expected.

I am going to take a break here. The title of this caught my attention while I was looking for no regret learning yesterday, so I'll watch it later. I guess I'll be going through Lanctot's references earlier than I thought I would. Maybe I will get somewhere today."