"1pm. Let me watch that video.
1:15pm. https://youtu.be/49PHeGZzttM?t=562

This video has been quite good so far. It introduces no-regret learning quite well.

1:20pm. https://youtu.be/49PHeGZzttM?t=844
"There are modern connections to Black-Scholes. In particular it is possible to derive the Black-Scholes pricing formula without any probabilistic assumptions whatsoever by essentially viewing the pricing process as a game between a price setter and an arbitrageur both of whom are playing a no regret algorithm. This was done in the last few years or so."

The video is from 2017.

1:40pm. https://youtu.be/49PHeGZzttM?t=1808
"
* Could ask for no regret to best strategy in risk-adjusted metrics:
  - Sharpe ratio: mean(returns)/std(returns)
  - mean - variance: mean(returns) - std(returns)
* Strong negative results
  - no-regret provably impossible
  - lower bounds on competitive ratio for any algorithm
"

2pm. https://youtu.be/49PHeGZzttM?t=2713

Maybe I will take a look at the first paper for this.

https://www.cs.princeton.edu/~arora/pubs/MWsurvey.pdf
The Multiplicative Weights Update Method - a Meta Algorithm and Applications

2:15pm. Let me take a break. Then I'll read this paper. Then come the references by Lanctot.

3pm. That took longer than I thought it would.

Let me get back on track. The paper.

...Ah, yeah. What was the softmax update again? Isn't it exactly the same as the outcome sampling one?

`ret {in={prob label}; map=inl {prob label} -> (prob - label) / temp} (adjoint input)`

So it is `probability - label`.

```
let r = if i = action_index then W * (1.0 - proposal_distribution.[i]) else -W
```

Yeah, I am a fool. This is exactly that...no wait it is not. If it were it would be something like...

```
let r = if i = action_index then W * (1.0 - proposal_distribution.[i]) else -W * proposal_distribution.[i]
```

3:35pm. Thunder. The weather is turning bad again. Let me turn off. I'll resume later.

6:20pm. Damn, that took a while to pass.

At the very least I read that paper for the first time. It is pretty mindblowing that the multiplicative updates can be used to approximately solve various LP and NP-Hard problems. I could not understand most of it, but I can tell that this paper is to be kept in mind. Solving constraints is a major function of biological brains, and I am on the lookout for how that could be plausibly be done on the computer.

CFR and now this...I've been living in a very small box this whole time it seems. The number of ways multiplicative updates can work is staggering. I need to study this. I need to master this. If I can do it, I will be able to make human level agents by the time I reach the end of this path. It will be a new beginning.

I have my work cut out for me. I am just going to keep plugging away at it.

But by now I am starting to be convinced - vanilla RL is not going to cut it. CFR is essentially a two phase (slow/fast policy) actor-critic method, but while I am going to need it, unfortunately I have no idea how I am going to use it together with neural nets. Or how I am going to generalize it to regular RL from simple, finite games like poker. I do not know how to tie the two fields together.

I hope there are some good ideas in the variance reduction papers.

Anyway, I got a reply from Lanctot to the softmax question. I really do appreciate him being so helpful.

```
> Let me just ask a quick question about the softmax. The PRG book mentioned exponential averaging, which reminded me that I've yet to see any mention of that in any of the CFR papers that I've read. I am guessing that was tried and abandoned in favor of regret matching. Was that the case?

Yes, the exponential algorithms are more common in the online learning literature because they have better theoretical bounds (and I think easier to analyze), and they are just very commonly used.

In CFR there are three very good reason to use regret matching: (i) RM just plain works better in practice, (ii) RM does not have a parameter that requires tuning (i.e. learning rate, step size), and (iii) RM gives you policies where actions have probability zero, which means you can apply some simple pruning in the sequential tree setting since the reach probabilities go to zero and you can skip the subtrees because the updates are zero). If you're curious how these come in, take a look at the CFR implementation in OpenSpiel.
```

These are all good points. I'd add that regret matching is more biologically plausible as well, since it does not require computing those exponentials like in softmax.

6:35pm. Let me close with a quote from the paper.

> We feel that our meta algorithm and its analysis are simple and useful enough that they should be viewed as a basic tool taught to all algorithms students together with divide-and-conquer, dynamic programming, random sampling, and the like.

I agree. They presented more than enough evidence to convince me that this is worth studying.

By the end of the year, I should be able to gain something significant from this."