"10:05am. https://www.reddit.com/r/ProgrammingLanguages/comments/dccrgb/microsoft_lean_theorem_prover/ Future of Mathematics?
I ended up watching all of this yesterday which is the reason why I went to bed past 12am. It is quite an interesting talk.

Let me read Solo Leveling and then I will start. My long Legendary XCOM campaign is finally drawing to a close. The Chosen are dead, the research tree is nearly exhausted and I am starting to get bored with the missions thrown my way. But wow, it took me a while to get to this point and every day was heaven.

The anime season is done too and I am going to have to pick some new shows to watch.

10:15am. Well, let me chill a while and then I will start. Today on the menu are references by Lanctot.

10:30am. Let me start.

10:45am. https://arxiv.org/abs/1810.09026
Actor-Critic Policy Optimization in Partially Observable Multiagent Environments

Lanctot is the second author on this paper.

"There are two main problems with (MC)CFR methods: (i) significant variance is introduced by sampling (off-policy) since quantities are divided by reach probabilities, (ii) there is no generalization across states except through expert abstractions and/or forward simulation with a perfect model. We show that actor-critics address both problems and that they are a form of on-policy MCCFR."

This is actually looks quite promising. Right now I am busting my head to try and figure out how I could port the CFR algorithm to use function approximation in a principled manner.

11:10am. 8/28. Interesting how much better QPG and RPG are than A2C. The main difference seem to be having state action baselines (which are then integrated over.) I was going to use just the state baselines so this is a good indicator that I should do it differently.

Very well.

11:15pm. "Our experiments show that these actor-critic algorithms converge to approximate Nash equilibria in commonly-used benchmark Poker domains with rates similar to or better than baseline model-free algorithms for zero-sum games. However, they may be easier to implement, and do not require storing a large memory of transitions. Furthermore, the current policy of some variants do significantly better than the baselines (including the average policy of NFSP) when evaluated against fixed bots. Of the actor-critic variants, RPG and QPG seem to outperform RMPG in our experiments."

This is great. So PG do work on poker. Nice.

11:25am. "One main crticial point is that temporal difference bootstrapping from values recursively is not possible as the Markov property does not hold in general for information states in partially-observable multiagent environments: the optimal policy πi(s, a) at some state st does generally depend on the policies at other information states."

Note the typo. `crticial` should be `critical`.

11:35am. "Let us recall the updated we consider in this paper"

`updated` should be `updates`.

11:50am. "Tabular policies are represents in behavioral strategy form"

`represents` should be `represented`.

`sum of partial derivates` should be `sum of partial derivatives`.

12:10pm. Done with the CFR actor critic paper. This was actually just what I needed. Great. The most important thing is that it converges on these partially observable multi player games. I was really worrying how I would be able to adapt CFR to this, but this takes care of it all. To think that AC algorithms with some slight tweaks will suffice.

12:15pm. There are some details I still do not get, but they are trivial in comparison to my previous worries.

After reading the VR-MCCFR paper, I've been hypothesizing that integrating state action pairs to get the baseline might be better than just using a state baseline in an AC system and this answered that decisively in a positive manner. I can easily combine this with Zap too. I will get all of this to work without a doubt.

12:20pm. Let me take a break here. Things are going well for me."