"8:50am. I am up. Well, before I can start, let me do that side chore I mentioned yesterday. I am acting as tech support for an acquaintance of mine. You'd think it would be easy to install Windows on some laptop, but for some reason none of the drivers work and I've had to force them through the Device Manager. If it was just that then that would be it, but unfortunately out of everything, the touchpad drivers fail to work. So what am I going to do now?
See if some setting is not set in BIOS, try to figure out why the device is not showing in Device Manager and scour the net after that. If that fails, I am not sure. I've gone through this plenty of times and always pulled through in the end, but who knows. I am also wondering if the 4.5 year old Windows 10 ISO is the problem and whether I should get a newer one. I've already run into issues like this once before.

At any rate, I'll resume my studies when I am done with this. It is going to take time out of my day.

9:15am. Uf, I have no idea. The touchpad simply cannot be found.

9:20am. Nope. The device is just invisible to the OS. Not much I can do about that, but get a newer one. Let me give that a try.

10:05am. While the ISO is downloading, I can get some reading in thankfully.

10:15am. https://arxiv.org/abs/1709.00503
Mean Actor Critic

10:50am. The MAC paper is interesting. I did not think it would update based on actions that weren't executed.

Wait, wait. Is is updating on the future state's values or the current ones? Well, `Q(s,a)` is the value of the future state given an action. But to me there is something fundamentally different about `V(s')` and `Q(s,a)`. The later is an estimate from the current point, while the former is literally taken from one step ahead.

The interesting part about using `Q(s,a)` is that I do not need to wait for the next action in order to make an update to the actor.

Interesting. This completely slipped my mind while I was reading that paper by Lanctot. I really thought it was suspicious the way that `Q` was used in the paper.

Always rather than using `Q(s,a)` I would just grab the value from the next state directly. I never tried the advantage variant of the AC.

11:30am. I am reading the paper by Lanctot again. As expected they are in fact using Monte Carlo to train the critic. For the actor, just as they say, they are using the MAC update. Interesting. Interesting...

At this point, the evidence is so strong that I should really doubt the claim that bootstrapping would not work for the critic. Everything is set for me to combine this with Zap. Since Zap updates work for Q learning, I will have no trouble making it work with this, once I used discrepancy reduction to make sure the base is stable.

I only really need discrepancy reduction to separate the head from the body of the network. I could do it layer by layer as I wrote in the draft, but I do not really need it there. Backprop works fine in SL regimes.

11:35am. No wait...shit. I won't be able to do the update like I want. Not unless I separate the inputs and actions into separate bins.

12:10pm. Ahhhhhhhh...

Change of plans. I am going to forget about Zap. There is no way to make it work neatly. Having actions in addition to state would really wreck having to calculate that steady state matrix.

12:15pm. Forget that for a while. I am just going to focus on making it work normally.

Well, maybe I will not need it given all the sharing done in deep nets.

12:20pm. Now I am thinking about fixed horizon methods.

12:25pm. No, the ideas I have won't work.

Agh, forget it.

One thing at a time. Forget higher order methods. Forget all this nonsense. I'll figure it out eventually. I just have to take it one thing at a time.

I will make new policy. Stop wasting time thinking about higher order methods like Zap until I've exhausted all the stuff in these papers in my testing.

12:30pm. I've done a lot of wankery in 2018 trying to put various speculative things together. This time, I should just do what works.

...I mean, it is confusing why these new ACs would work while the vanilla A2C fails given that they are not that much different from each other. Maybe A2C+Zap could be made to work. But it is not really my job anymore.

I am going to stop being that mad kind of researcher.

12:40pm. New goal. I need to test those actor critic architectures in a tabular setting. Does the A2C fail specifically because of what it is, or is there some other problem?

This is what I should have done. Before I start thinking about how I am going to shove Zap into the top layer, what I need are sanity checks in the tabular case.

If it turns out that QPG and RPG work in the tabular case, but A2C does not, then I will have my answer regarding what I need to do. If A2C does in fact converge to the Nash equilibrium, then I can think of how to apply Zap to it in hopes of fixing its performance.

12:45pm. Let me finally have breakfast. It has been long overdue."