"1:40pm. https://www.youtube.com/watch?v=I_CCQ-AC5gM The Blessing and the Curse of Multiplicative Updates
I'll watch this later.

1:45pm. I have to do some chores. Let me go outside for a bit.

4:40pm. I am back. Now before I throw away Zap for good, let me take a look at the Nelsa and Polsa from the successor paper again.

Could I adapt those for my use case?

5pm. Hmmm, NeSa might be usable when the Q function has multiple actions.

5:20pm. But what if the update is skipped on some turn.

Really, the thing that really bugs me here is suppose I had an input of size 128, and 128 actions. Does that mean I would need...`128*128=16384`...

Does that mean I would need 16384*16384 steady state matrix even in NeSa's case? Well, no since it is one action, no doubt I would only need 128*128. But what do I do about the cases where the action was not chosen? What do I do about the differential theta term? What?

In the original Zap paper, what is the example they are testing it on?

5:30pm. https://arxiv.org/abs/1707.03770
Fastest Convergence for Q-learning

23/46.

"Consider first a simple stochastic-shortest-path problem. The state space X = {1, . . . , 6} coincides with the six nodes on the un-directed graph shown in Fig. 2. The action space U = {ex,x0}, x, x0 ∈ X, consists of all feasible edges along which an agent can travel, including each “self-loop”, u = ex,x. The number of state-action pairs for this example coincides with the number of nodes plus twice the number of edges: d = 18.

The controlled transition matrix is defined as follows: If Xn = x ∈ X, and Un = ex,x0 ∈ U, then Xn+1 = x 0 with probability 0.8, and with probability 0.2, the next state is randomly chosen between all neighboring nodes. The goal is to reach the state x ∗ = 6 and maximize the time spent there. This is modeled through a discounted-reward optimality criterion with discount factor β ∈ (0, 1). The one-step reward is defined..."

Let me skip ahead a bit.

"Based on this analysis or on Thm. 3.3, it follows that the asymptotic covariance is not finite for the standard Watkins’ algorithm with αn ≡ 1/n. In simulations it was found that the parameter estimates are not close to θ ∗ even after many millions of samples. This is illustrated for the case β = 0.8 in Fig. 5, which shows a histogram of 103 estimates of θn(15) with n = 106 (other entries showed similar behavior).

It was found that the algorithm performed very poorly in practice for any scalar gain algorithm. For example, more than half of the 103 experiments using β = 0.8 and g = 70 resulted in values of θn(15) exceeding θ ∗ (15) by 104 (with θ ∗ (15) ≈ 500), even with n = 106 . The algorithm performed well with the introduction of projection in the case β = 0.8. With β = 0.99, the performance was unacceptable for any scalar gain, even with projection..."

5:35pm. I can take this at face value which I did a year ago.

But at this point, I really have to wonder - what exploration method were they using? This is a huge thing in RL.

If they were using zero init - then it is perfectly plausible that the thing might get stuck.

If they were doing optimistic init - then the thing will do the right thing and converge in less than 10k samples guaranteed.

I should have pushed the author about this back in 2018.

5:50pm. I need to stop being so naive and stop taking things on its head. What really matters in RL is exploration. I do not need to bend over to accommodate second order methods.

5:55pm. Alternatively, maybe they did use optimistic init, but it did not matter. What they were really looking for are optimal values, not the optimal policy.

This might be a good lesson for me. The tabular thing might easily find the optimal policy, but not so much the optimal parameter values for the whole MDP.

6pm. It does not matter.

6:05pm. Though I spent so much time on it, I am going to get rid of KFAC and Zap from my arsenal. They are just dead weight. I do not need them. There are better substitutes. Yes, KFAC helps, but I can get the same effect by running the net a few more times.

6:10pm. Absolutely, I am going to do things right this time around. Forget GPUs, forget Zap, forget differentiable plasticity, forget RUDDER, forget KFAC.

Remember correct experimental procedures. Remember good engineering. Remember good science.

And learn the power of simple multiplicative updates.

Along with reversibility, they are the ones that will allow me to crack what I want.

I vastly overrated exotic things, and vastly underrated the insight one can draw from simple things scaled up. I have to apologize to tabular RL.

6:20pm. I swear, I am going to ask Lanctot about how those ACs act in the tabular case tomorrow. Also ask him what is meant by vanilla RL not converging to a Nash equilibrium in zero sum games.

I meant to do that earlier in the day, but chores, reading and me obsessing over old things are getting in the way.

I really want to just continue reading the MAS book, but I have to call it a day here. It is a principle that I stop after 6pm.

The principle of laziness. This principle is not so much about brute forcing infinities as it is about keeping my stress levels manageable."