"9:35am. I am up. Today what is on the menu is that game theory course.
9:40am. https://mangadex.org/title/41946/genocide-online

This got a manga. Well, let me read this and then I will start.

10:10am. Ok, let me start. I need to get the game theory stuff out of the way.

10:40am. https://youtu.be/7jBf5fzGBlk?list=PLeY-lFPWgBThlAF5VFWlOWy5zqhkXcCqz&t=339

I am realizing something. I thought that I had the general framework for computing Nash equilibria in CFR, but I am not sure that it would converge to a randomized strategy in battle of the sexes which is not a zero sum game.

Rather I think it would converge to one of the slots.

...Actually I am not sure. Maybe if you inited it that way exactly it would be true. But in random init...

It is just so difficult to reason out what an optimization algorithm would do directly.

12pm. https://www.youtube.com/watch?v=bmL8UxMINyo&list=PLeY-lFPWgBTiXWuvK2ud2ZJySQ6pilENH

I'll leave this for later.

Let me have breakfast.

1:40pm. Uh, I forgot to commit earlier. Well, nevermind that. What I am going to do now are chores. In an hour or so I'll get back to watching a few more lectures.

Yeah, no way am I going through 6 lectures in a day, but 3 seem doable."