"12:50pm. Done with breakfast. Let me go do the chores and then I will dive into the paper by Heinrich. There is so much that I want to do, and yet I am getting bogged down in these useless algorithms. Once again, the fact that Lanctot cannot clearly explain strategy averaging makes me think that maybe he himself is not too familiar with the specific details of what I am asking him.
I was hoping the books and the lectures would make things more clear, but that effort was a flop.

3:45pm. I am back.

5pm. Done with the bath.

Literally the only thing on my mind is how to make Zap work it seems. I am considering a 3 phase algorithm. Train V using Zap, train Q using KFAC and then the actor based on that.

I am absolutely obsessed about this. Making the value function stable is the most important part of deep RL. I keep looking for any kind of way to make use of Zap for that reason.

I do not want to use Monte Carlo if I can help it.

But I can't really use Zap to train the Q function. So the answer is - train the V function and train the Q off it in a supervised fashion.

5:15pm. I've been constantly toying with the idea of making the Q function V + value of the action, but I should avoid that sort of thing. Rather I should train Q directly. That way will be the most stable and with the least amount of hassle.

The other best alternative if I really want to avoid bootstraping which is reasonable in short games would be to train the Q function using MC values.

But the value function is the ultimate connector. Without it, I won't be able to bring down the variance to its limit. I want to bring all the tricks that I've learned together into one whole.

Discrepancy reduction, KFAC and Zap.

5:25pm. Agh, forget that. Let me use the last half hour that is left in the day to read that paper.

5:35pm. 4/9.

"The extensive-form process ...  is realization-equivalent to a generalised weakened fictitious play in the normal-form and therefore the average strategy profile converges to a Nash equilibrium in all games with the fictitious play property."

I keep seeing this over and over, but I am yet to see any explanation of why this actually makes sense. Why is that equation equivalent to the normal form thing?

The equation in this paper is particularly convoluted.

5:50pm. Ok, let's say that studying game theory for a bit was interesting. But I had enough of it and these papers. Tomorrow I am going back to Lanctot's thesis and resuming work on outcome sampling. It is time that I resume. The things I cannot figure out do not matter.

It seems obvious that RL will remain being convoluted despite my best efforts and that I will have to make do with what I have.

6pm. This is so annoying. It is really taking me way too long to do something that should have been trivial, but I haven't touched ML since 2018. I could not beat the odds, but I think thanks to that experience that I finally have what it takes to make a successful attempt.

There just isn't that much to do.

In the end there will just be ACs with some tricks. I will take care of that, take care of the games, and then take care of those online gambling dens.

Once I have these basic pieces set, only then it will become clear how to put them together in a hierarchy.

The advanced agents capable of doing metalearning are only a step beyond the basic ones.

But the basics have to be done properly.

https://www.youtube.com/watch?v=I_CCQ-AC5gM
The Blessing and the Curse of Multiplicative Updates

And the very basics are always these multiplicative updates. I am going to get deeply familiar with them starting from here on out. Actor Critics are merely them + some variance reduction.

If one piece is missing, such as discrepancy reduction, then the whole construction will fall apart.

Really, with it, everything should be there.

By the end of 2020 and much sooner, I will know for sure whether the basics are there from real experience.

At that point, the focus will change from eating weeds to gazing at the stars. I will be able to start thinking about not just how to make the agents work, but how to make them good. How to make them capable of winning in not just carefully designed playgrounds like poker, but in games that capture the essence of the real world.

The level after that is the Singularity. The algorithms and the hardware will be there, and then it will be time for me to really show my true worth.

6:15pm. Well, the first step is to stop eating weeds. I want some meat!

I am starving for it!

Since I am done with studies, the next year will be a whirlwind of programming!

I will do what I could not in 2018!"